requirements.txt
ppython-telegram-bot==20.7
python-dotenv==1.0.0